1 - Implementation basic client-server connection
- improved tests and adjusted naming of test class
- fixed java version in build.yml
- enabled Jacoco HTML reports for debugging
