+++ build.gradle.kts
@@ -58,5 +58,6 @@ tasks.jacocoTestReport {
 
     reports {
         xml.required.set(true)
+        html.required.set(true)
     }
 }
