+++ src/test/java/com/example/cataniaunited/api/GameWebSocketTest.java
@@ -1,0 +1,114 @@
+package com.example.cataniaunited.api;
+
+import io.quarkus.test.common.http.TestHTTPResource;
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.websockets.next.BasicWebSocketConnector;
+import io.quarkus.websockets.next.OpenConnections;
+import jakarta.inject.Inject;
+import org.junit.jupiter.api.Test;
+
+import java.net.URI;
+import java.time.Duration;
+import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+@QuarkusTest
+public class GameWebSocketTest {
+
+    @TestHTTPResource
+    URI SERVER_URI;
+
+    @Inject
+    OpenConnections connections;
+
+    @Test
+    void testWebSocketOnOpen() throws InterruptedException {
+        List<String> receivedMessages = new ArrayList<>();
+        CountDownLatch messageLatch = new CountDownLatch(1);
+        var openConnections = connections.listAll().size();
+        BasicWebSocketConnector.create()
+                .baseUri(SERVER_URI)
+                .path("/game")
+                .onTextMessage((connection, message) -> {
+                    receivedMessages.add(message);
+                    messageLatch.countDown();
+                })
+                .connectAndAwait();
+
+        // Wait up to 5 seconds for both messages to arrive
+        boolean allMessagesReceived = messageLatch.await(5, TimeUnit.SECONDS);
+
+        assertTrue(allMessagesReceived, "Not all messages were received in time!");
+        assertEquals("Connection successful", receivedMessages.getFirst());
+        assertEquals(openConnections + 1, connections.listAll().size());
+    }
+
+    @Test
+    void testWebSocketSendMessage() throws InterruptedException {
+        String sentMessage = "Hello World!";
+        List<String> receivedMessages = new ArrayList<>();
+        // Expecting 2 messages
+        CountDownLatch messageLatch = new CountDownLatch(2);
+
+        var webSocketClientConnection = BasicWebSocketConnector.create()
+                .baseUri(SERVER_URI)
+                .path("/game")
+                .onTextMessage((connection, message) -> {
+                    receivedMessages.add(message);
+                    messageLatch.countDown();  // Decrease latch count when a message arrives
+                })
+                .connectAndAwait();
+
+        // Send message
+        webSocketClientConnection.sendTextAndAwait(sentMessage);
+
+        // Wait up to 5 seconds for both messages to arrive
+        boolean allMessagesReceived = messageLatch.await(5, TimeUnit.SECONDS);
+
+        assertTrue(allMessagesReceived, "Not all messages were received in time!");
+        assertEquals(2, receivedMessages.size());
+        assertEquals(sentMessage, receivedMessages.getLast());
+    }
+
+    @Test
+    void testWebSocketOnClose() throws InterruptedException {
+        List<String> receivedMessages = new ArrayList<>();
+        // Expecting 2 messages
+        CountDownLatch messageLatch = new CountDownLatch(2);
+
+        var clientToClose = BasicWebSocketConnector.create()
+                .baseUri(SERVER_URI)
+                .path("/game")
+                .connectAndAwait();
+
+        BasicWebSocketConnector.create()
+                .baseUri(SERVER_URI)
+                .path("/game")
+                .onTextMessage((connection, message) -> {
+                    receivedMessages.add(message);
+                    messageLatch.countDown();  // Decrease latch count when a message arrives
+                })
+                .connectAndAwait();
+
+        assertEquals(2, connections.listAll().size());
+
+        // Send message
+        clientToClose.close().await().atMost(Duration.of(5, ChronoUnit.SECONDS));
+
+        // Wait up to 5 seconds for both messages to arrive
+        boolean allMessagesReceived = messageLatch.await(5, TimeUnit.SECONDS);
+
+        //Check if client disconnected
+        assertEquals(1, connections.listAll().size());
+        assertTrue(allMessagesReceived, "Not all messages were received in time!");
+        assertEquals(2, receivedMessages.size());
+        assertEquals("Client disconnected", receivedMessages.getLast());
+    }
+
+}
